WIP refactor to push client to finish signUp after Google auth

+++ migrations/20220326214403_create_user_post_and_chat_tables.js
@@ -2,7 +2,7 @@ exports.up = function (knex) {
   return knex.schema
     .createTable('users', (table) => {
       table.increments('id').primary();
-      table.integer('google_id').notNullable();
+      table.bigInteger('google_id').notNullable();
       table.string('avatar_url').notNullable();
       table.string('username').notNullable();
       table.string('first_name').notNullable();
